requirements.txt
streamlit
pillow
pycryptodome